# Producer_SPI/src/main/java/mk/ukim/finki/thesis/spi/model/Product.java
package mk.ukim.finki.thesis.spi.model;

import lombok.Builder;
import lombok.Getter;

@Getter
@Builder
public class Product implements UserActivity {
  private Long id;
  private String name;
  private double price;
  private String category;
}