# Producer_Web/src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  title: E-Shop API
  description: API for an e-shop allowing users to view products, add them to the cart, cancel orders, and search for products.
  version: 1.0.0
paths:
  /product/{productId}:
    get:
      summary: View a product
      description: Retrieve detailed information about a specific product.
      parameters:
        - name: productId
          in: path
          required: true
          description: The ID of the product to retrieve.
          schema:
            type: number
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
  /cart:
    post:
      summary: Add a product to the cart
      description: Add a specified product to the user's shopping cart.
      requestBody:
        description: Details of the product to add to the cart
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: number
                quantity:
                  type: integer
      responses:
        '200':
          description: Product added to cart successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        '400':
          description: Invalid input
  /order/cancel:
    post:
      summary: Cancel the order process
      description: Cancel the current order process.
      requestBody:
        description: Request to cancel the order
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId:
                  type: number
      responses:
        '200':
          description: Order process cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCancellation'
        '404':
          description: Order not found
  /search:
    get:
      summary: Search for products
      description: Search for products using a search string.
      parameters:
        - name: query
          in: query
          required: true
          description: The search string to use for finding products.
          schema:
            type: string
        - name: userId
          in: query
          required: true
          description: The user id of the user performing the search.
          schema:
            type: number
      responses:
        '200':
          description: Search results returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '400':
          description: Invalid search string
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
    CartItem:
      type: object
      properties:
        cartId:
          type: number
        productId:
          type: number
        quantity:
          type: integer
    OrderCancellation:
      type: object
      properties:
        cartId:
          type: number
        userId:
          type: number
        productIds:
          type: array
          items:
            type: number
    SearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Product'